NODE-937: Clean up
